feat: add CIntroSession to build with VK_ key definitions and header fixes
- Add VK_PAUSE, VK_F1, VK_F2, VK_F11 to Kernel.h
- Add TSUISettings.h include to GameSettings.h (for CExtensionListMap)
- Add TSUIHUD.h include to GameSessionDisplays.h (for SHUDPaintCtx, IHUDPainter)
- Fix CMenuDisplay::SOptions struct (removed default initializers)
- Fix pointer cast in CTranscendencePlayer::CreateGlobalRef (int to intptr_t)
- Simplified AppCore.h by removing duplicate declarations and moving SAppState definition inline
- Keep CIntroSession in build (was added then removed to test incremental)

+++ Transcendence/Transcendence/Transcendence.h
@@ -798,7 +798,7 @@ class CTranscendencePlayer : public IPlayerController
 
 		//	IPlayerController interface
 
-		virtual ICCItem *CreateGlobalRef (CCodeChain &CC) override { return CC.CreateInteger((int)m_pPlayer); }
+		virtual ICCItem *CreateGlobalRef (CCodeChain &CC) override { return CC.CreateInteger((intptr_t)m_pPlayer); }
 		virtual CPlayerGameStats *GetGameStats (void) const override { return &m_pPlayer->GetGameStats(); }
 		virtual GenomeTypes GetGenome (void) const override;
 		virtual DWORD GetLastWarningTick () const override { return m_EphemeralState.dwLastWarningTick; };
